requirements.txt
-i https://pypi.org/simple
boto3==1.28.20
botocore==1.31.20
jmespath==1.0.1; python_version >= '3.7'
pyjwt==2.8.0
pynamodb==5.5.0
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2023.3
s3transfer==0.6.1; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.16; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
